Adding field breakout data sets
