+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include us_birth_data/us_birth_data.parquet
+include us_birth_data/*.parquet
